add test for tsx
